+++ app/root.tsx
@@ -4,11 +4,16 @@ import {
   Meta,
   Outlet,
   Scripts,
-  ScrollRestoration
+  ScrollRestoration,
+  isRouteErrorResponse,
+  useRouteError
 } from "@remix-run/react"
+import { HttpStatusCode } from "axios"
+import { BiSolidError } from "react-icons/bi"
 
 import createMeta from "~/libs/meta"
 
+import TopButton from "./components/top-button"
 import styles from "./styles/app.css"
 
 export function meta() {
@@ -31,6 +36,38 @@ export default function App() {
         <ScrollRestoration />
         <Scripts />
         <LiveReload />
+      </body>
+    </html>
+  )
+}
+
+export function ErrorBoundary() {
+  const error = useRouteError()
+
+  const status = isRouteErrorResponse(error)
+    ? error.status
+    : HttpStatusCode.InternalServerError
+
+  const message = isRouteErrorResponse(error)
+    ? error.statusText
+    : "Unknown error"
+
+  return (
+    <html>
+      <head>
+        <title>ERROR!</title>
+        <Meta />
+        <Links />
+      </head>
+      <body className="flex min-h-screen w-screen items-center justify-center">
+        <BiSolidError className="mr-6 text-6xl" />
+        <div className="stat w-fit border-l-2">
+          <div className="stat-title">ERROR</div>
+          <div className="stat-value">{status}</div>
+          <div className="stat-desc">{message}</div>
+        </div>
+        <TopButton />
+        <Scripts />
       </body>
     </html>
   )
